Change Recording length type from int to unsigned long for better duration handling

+++ include/Models.h
@@ -4,7 +4,7 @@
 struct Recording {
     String username;
     time_t timestamp;
-    int length;
+    unsigned long length;
 };
 
 struct Credentials {
